go-to-gym-platform/backend/core_api/auth/serializers.py
from rest_framework import serializers


class ForgotPasswordSerializer(serializers.Serializer):
    email = serializers.EmailField()


class ResetPasswordSerializer(serializers.Serializer):
    new_password = serializers.CharField(write_only=True)
    confirm_password = serializers.CharField(write_only=True)

    def validate_new_password(self, value):
        if len(value) < 8 or not any(c.isupper() for c in value) or not any(c.isdigit() for c in value):
            raise serializers.ValidationError(
                'La contraseña debe tener al menos 8 caracteres, una mayúscula y un número'
            )
        return value

    def validate(self, attrs):
        if attrs['new_password'] != attrs['confirm_password']:
            raise serializers.ValidationError({'confirm_password': 'Las contraseñas no coinciden'})
        return attrs
