requirements.txt
Django>=5.2
wagtail>=6.0
django-crispy-forms
crispy-tailwind
django-widget-tweaks
django-taggit
django-modelcluster
Pillow
mysqlclient
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary

python-oqs
cryptography>=42.0.0
